Break up instructor and dashboard
