+++ src/components/InstructorLessonsWidget/index.js
