+++ src/components/InstructorRevenueWidget/utils/removeRevenueMonth/index.js
