+++ src/components/InstructorRevenueWidget/utils/removeRevenueMonth/index.test.js
