+++ src/components/InstructorRevenueWidget/utils/totalRevenue/index.js
