+++ src/components/InstructorRevenueWidget/utils/totalRevenue/index.test.js
