+++ src/screens/Dashboard/components/GetPublishedWidget/components/Checklist/components/MoreInfo.js
