+++ src/screens/Dashboard/components/GetPublishedWidget/components/Checklist/index.js
