+++ src/screens/Dashboard/components/GetPublishedWidget/utils/isStepComplete/index.js
