+++ src/screens/Dashboard/components/GetPublishedWidget/utils/isStepComplete/index.test.js
