+++ src/screens/Dashboard/components/ProposeWidget/index.js
